Backend: Default to localhost port 4000 for the database
This default configuration is only useful when executing the backend locally through the ide through the
executable jar file built by Spring Boot and can be easily overriden by specifying .
The main consideration there is that a docker container with the database is running which exposes its mysql
database on port 4000 of the host. Mainly so it does not interfere with a mysql database thats possibly
already running on the host with the default post (e.g 3306).