Client: Allow configuration of environemnts
This patch adds the ability to configure environments, and includes two preconfigured
environments, one for production and one for development.

Run 'npm run config_production' or 'npm run config_development' to configure either
of the environments.

The environments are used to set the location url of the REST API of the backend.

This patch also integrates with maven, so when packaging the app the production
environment will be selected