[Client] Implement RoleManager, refactor role initialization
This removes our simple role and user initialization from cookies and implements a RoleManager that tracks the
allowed security roles of the current user. We switch to using symbolic role names from the 'Role' class.

Switching roles has been simplified and only sets the current role in the UserService and forwards to the
correct route.

Login and Logout works correctly now but setting and resetting the required states in the UserService.

The logging during route changes is improved.

The Navigation at the top correctly handles the new route system now.