Enumeration: Fix enumeration across subtrees
This fixes the way entries are enumerated across
exercise subtrees. When we need to recurse into
the subtree of an Exercise to enumerate content
for correctly or incorrectly answered exercises,
then we need a way to return to the original tree
when enumeration has finished.

In order to return from subtrees to the parent
tree, we change the way the client and backend
track enumeration ids. Rather than using a single
property 'nextId' we now pass a list of 'nextIds'
from backend to client.

When the client requests new content it will send
the list of 'nextIds' back to the backend, which
will use it during the enumerations.