[Backend] Database: Switch to flyway
With this patch we switch schema creation to flyway. That means that in the future, any
change to the database must be handled with a migration script, which has to be placed
into the resources/db/migration folder and which must follow the flyway naming schema.

Starting with this patch the database will no longer be reset during application
restart.