[Backend] UserController: Refactor for more reusability
We want to make it possible to bootstrap the initial user without the need to login. Such that, when
the application is started for the first time with an empty database the admin can create the first
user without any credentials.

This patch prepares for that by refactoring the createUser method into public method
named 'createUserFromDto', which is does not export a rest endpoint and which can be reused
by other controllers. Most of the logic is also moved to the UserService.