[Client] Make profile shared between all roles.
This patch removes the teacher profile and moves the
student profile to the user/ module. The profile
is now a shared profile between all three roles.

Additionally the profile menu entry is now also shown
in the user menu if any of the roles is active and not
only for the student role.