[Backend] EntryHistory: Add user reference to EntryHistory and update schema
This patch is a fix for the previous commit 5fb2c789e7bd37e52b88cebe5a168b3e20d0542c
which was reverted in 78a8e32f72ccb747ef55a3635e2afd6c102a4d51 since it was missing a
statement to update existing in the table.

This patch adds a new column named 'user_id' to the 'entry_history' table.

It also adds the update statement that was missing in the original patch. This
updates the entry_history table after the new column was added and
populates the new column with the right values from the 'course_history' table.

Without this update statement the last statement, which adds a foreign key, would fail
if the table had previously contained any data. In other words the migration would only be
successful if the entry_history table was empty. The update statement makes sure that adding
the foreign key works on a production database with existing data.