[Client] Docker: Fix entrypoint and template installation
The directive to add the default.template had a typo
and tried to add the default.conf which we had
already deleted.

Rather then using the 'RUN' directive which runs
at image creation time, we need to use the
'ENTRYPOINT' directive.

An important part in the entrypoint directive is
the 'exec' command after && and before nginx. This
ensures that nginx is not forked as a child process
of bash but replaces the bash process. As a result it
will be able to receive unix signals such as SIGTERM.

Also note the addition of the string '$NGINX_HOST'
(with single quotes). This makes sure that envsubst
command will only replace the variable in this string
(e.g. $NGINX_HOST) and not any text sequence starting with
a $ sign. This is important since the nginx configuration
file contains some $ prefixed strings which should
remain in the configuration file.