[Backend] CourseController: Add DELETE method for course
This method will not only delete the course object but traverse the whole course tree and delete all the attached
objects from this course. This inclused the complete user history for this course. The history of other courses
will not be affected. It also includes the whole entry tree attached to this course.