[backend] user import: if transaction logic
The transaction logic had an error, so transaction would not be rolled back
if the sending of the mail failed. Now the transaction is rolled back
such that the user is not added to the data base. Additionally,
the import is stopped and a status message is returned to the user.