[client] admin: fix pagination issues
This fixes a couple of small pagination issues. We now limit the number of page items
to 10. This revealed a small bug in the ui-bootstrap library, so we needed to
update the ui-bootstrap library to version 2.5.
We also actived boundary links to display first and last buttons in the
pagination and we enforce display of elipsis if there are more than
10 page items.