[backend] tests: correctly use test-database configuration
This change ensures that the test-database is correctly used instead of the
default database. The test database is expected to run under port 4001
while the default database runs under 4000.

Previously the app.properties file under test/resources was ignored,
by renaming it to follow the Spring convention (e.g. application-default.properties)
we ensure that Spring loads it automatically during tests.

Not that we cannot just name it application.properties because then it would
completely override the default application.properties file in the main/resources
classpath. However we only want to selectively override individual properties.

Since by default, the 'default' profile is active, we can name the
properties file application-default.properties to ensure that it
is loaded by Spring and also that it only selectively overrides
properties from the application.properties file.

A more detailed discussion about this issue can be found here:
https://stackoverflow.com/a/45021819/474034