[backend] Bump flyway version to 4.2.0
This is necessary so we can migrate to Spring Boot 2.

A release with this flyway version needs to be deployed to the
production server before a newer version with Spring Boot 2 can
be deployed. The reason is that Spring Boot 2 will use flyway
version 5 which does not support a migration path from flyway
version 3 which was the default in Spring Boot 1.x