[client] app: update current role on successful state change
This fixes an issue that prevented the current role to be
correctly set, when the user navigates directly to a certain
route rather then manually switching it though the
user menu.

We now switch the route directly in the transition function,
rather then in the switchRole function that is invoked
from the user-menu.